moved pages folder content to app folder
