+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,6 @@
+import { authOptions } from "@/utils/auth";
+import NextAuth from "next-auth/next";
+
+const handler = NextAuth(authOptions);
+
+export { handler as GET, handler as POST };
