+++ src/app/blog/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function page() {
+  return (
+    <div>
+      <h1>Coming Soon.</h1>
+    </div>
+  )
+}
