+++ src/app/global.scss
@@ -1,0 +1,49 @@
+body {
+  background-color: #ecf3f3;
+  width: 100vw;
+
+  @media screen and (max-width: 768px) {
+    background-color: white;
+    height: 100vh;
+    width: 100vw;
+  }
+}
+
+.particle {
+  position: absolute;
+  border-radius: 50%;
+  height: 0;
+}
+
+@for $i from 1 through 30 {
+  @keyframes particle-animation-#{$i} {
+    100% {
+      transform: translate3d(
+        (random(90) * 1vw),
+        (random(90) * 1vh),
+        (random(100) * 1px)
+      );
+    }
+  }
+
+  .particle:nth-child(#{$i}) {
+    animation: particle-animation-#{$i} 60s infinite;
+    $size: random(5) + 5 + px;
+    opacity: random(100) / 100;
+    height: $size;
+    width: $size;
+    animation-delay: -$i * 0.2s;
+    transform: translate3d(
+      (random(90) * 1vw),
+      (random(90) * 1vh),
+      (random(100) * 1px)
+    );
+    background: hsl(random(360), 70%, 50%);
+  }
+}
+
+@media screen and (max-width: 768px) {
+  .particle {
+    display: none;
+  }
+}
