--- src/pages/api/v1/projects/create/index.ts
@@ -1,51 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from 'next'
-import { Project } from '@/types'
-import prisma from '@/lib/prisma'
-import { getServerSession } from "next-auth/next"
-import { authOptions } from '@/pages/api/auth/[...nextauth]'
-
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse
-) {
-
-  const session = await getServerSession(req, res, authOptions)
-
-  try {
-    const { name, description, stacks, githubUrl, liveUrl, coverImgUrl, modalImgUrl, tag }: Project = req.body;
-    const project = {
-      name, description, stacks, githubUrl, liveUrl, coverImgUrl, modalImgUrl, tag,
-    };
-
-    if (!session?.user?.email) {
-      return res.status(401).json({
-        status: "error",
-        message: "You are not authorized to perform this action"
-      })
-
-    }
-
-    const result = await prisma.project.create({
-      data: {
-        ...project,
-        author: {
-          connect: {
-            email: session?.user?.email as string
-          }
-        }
-      }
-    });
-
-    res.status(200).json({
-      status: "success",
-      data: result,
-      message: "Project added successfully",
-    });
-  } catch (error) {
-    res.status(500).json({
-      status: "error",
-      error: error,
-      message: `Error adding project to database ${error}`,
-    });
-  }
-}
